Calculate size of side bar after drawing tables
There are situations where the size of the size of the side bar is
calculated before the size of the tables. The result is a side bar that
doesn't use all the page height. To address this issue the side bar
calculation has to be explicitly invoked from the table callback that is
executed after drawing.

Change-Id: I45b06f1dd6b94740adc7054bcb24872610e07aaa
Related-To: https://github.com/patternfly/patternfly/issues/121
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>