Use "colgroup" to set column widths
Currently the application uses the sets the width of columns using the
"width" attribute of the "th" element. But the data tables code uses
this to set the "width" attribute of the header cells, and as a result
the sorting indicator (the blue line above the column name) uses only a
part of the cell. For example, if the width of the column is 20% then
the sorting indicator uses only 20% of the cell. To avoid this glitch
this patch changes the application to use the table "colgroup" and "col"
elements to set the width.

Change-Id: Idcd3a328a5c280a957e1898cd7a58269ec3ee8b8
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>