Don't return value from function called by "each"
Currently the function used to highligh code called using the jQuery
"each" method returns a value, because it uses the "(..) => (expr)"
syntax. This isn't a very bad thing, but it isn't good either, as the
return value doesn't make sense. This patch modifies it so that it isn't
an expression, but a block, using the "(...) => {block}" syntax instead.

Change-Id: Ib170ecaacf03318d4651f2c7ea9b4c83785c9dce
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>