Return null link for null component
In some situations during the rendering of tables after specifying a
search criteria React will try to render a "Link" component passing
"null" as the "concept" parameter. This triggers an exception when
attempting to use the "name" property of the concept, and a side
effect is that the React router stops working completely. I wasn't
able to find why React is doing that, but I did find that returning
"null" from the function that builds the component in this case makes
React ignore it and continue. That is what this pach does, check the
"concept" parameter and return "null" if it is null.

Change-Id: I5450e6271055ed3da155b7721b41dde1aade81a6
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>