Use model names without modification
Currently the application takes the names of concepts from the model
description and transforms them. For example, the name "affinity_label"
type is transformed into "AffinityLabel". That is easier to read for
humans, but it isn't the name that needs to be used in the actual JSON
or XML documents sent to and recieved from the API. This patch changes
the application so that names aren't transformed, but used exactly as
they appear in the model.

Change-Id: I783c98427c432a86d90e74bda665af4ce7e73f48
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>