Introduce the 'Name' class
Currently the application manages the names of concepts as strings. This
isn't ideal, because names need to be displayed using different
strategies, depending on the kind of concept. For example, names of
types should be displayed using camel case, but names of struct members
should be displayed using snake case.  To make that easier that patch
introduces the 'Name' class, which stores and manages an array of words.
Later patches will take advantage of this class in order to actually
implement the different display strategies.

Change-Id: I38f40608b12ab28e45c49d5e9622c107869ab485
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>