Run source code highlight only once per document
Currently the source code highlight is executed once each time that an
HTML document is rendered. This introduces small delays, specially for
large documents. To avoid having that delay multiple times this patch
changes the application so that highlighting will run only the first
time. The result will be preserved and reused when the document needs to
be displayed again.

Change-Id: I068e15978bffcfaeabe93a4a82bea3331b74752a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>