Make name words mutable
Currently the when the list of words of a name is requested a copy is
created and returned, in order to avoid interference with the usages of
the same name in other places. This is expensive, and not needed in this
application, as it never modifies the names. In order to improve
performance in this area, this patch changes the code so that no copy is
performed.

Change-Id: I1f1841db8de2bff77b658f63c1abfe2933a5cc7a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>