Bundle JavaScript libraries
Currently the JavaScript libraries needed during runtime are loaded
using a 'script' tag in the 'index.html' file, so they are loaded from
an external content delivery network: cdnjs.com. That complicates using
the application in environments where there is no internet connection.
To improve that, this patch changes the application so that those
libraries are part of the bundle generated by webpack, so there is no
need for additional donwloads.

Note that the styles and fonts used by those libraries are still
donloaded from external servers. That will be addressed in later
patches.

Change-Id: I6b197d44ffda27699444ffced0d656591fb97e7d
Bug-Url: https://bugzilla.redhat.com/1404753
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>