Bundle styles and images
Currently the styles needed during runtime are loaded using a 'link'
tags, and they are loaded from an external content delivery network:
cdnjs.com. That complicates using the application in environments where
there is no internet connection. To improve that, this patch changes
the application so that those styles are part of the bundle generated by
webpack, so there is no need for additional donwloads.

Change-Id: I5fa1a90bed2d233c44b6f1d1f78f0381a320941e
Bug-Url: https://bugzilla.redhat.com/1404753
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>