Fix build issues
Currently the automated build of the project in Jenkins is failing
because the project is not yet prepared to build using the oVirt Node.js
modules. In addition the project will not be distributed using RPM. This
patch removes the RPM .spec files and changes the build process so that
it uses plain yarn to build and test, which should work without
problems.

Later patches will adjust the build so that it uses the oVirt Node.js
modules.

Change-Id: Ife3611ebe95cf5161771f10b99e2efd9ff609700
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>