Require ovirt-js-dependencies
Currently this application bundles all its dependencies. This means that
it doesn't reuse the depencencies provided by the 'ovirt-js-dependencies'
package. To improve that this patch modifies the application so that
it does not bundle the dependencies provided by that package, in
particular jQuery, Bootstrap and Patternfly. The application uses now
those libraries from a CDN, and the RPM packaging replaces the links to
that CDN with links to the same files provided by the
'ovirt-js-dependencies' package.

Change-Id: I876f6b1ff74a41b443efbbd9fd4af40bf27333fc
Bug-Url: https://bugzilla.redhat.com/1404753
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>